Move Monitor to util, add some documentation for it
